+++ nh_clinical/__init__.py
@@ -1,21 +1,20 @@
 # Part of NHClinical. See LICENSE file for full copyright and licensing details
 # -*- coding: utf-8 -*-
-from . import partner
-from . import user
-from .models import *
-from . import groups
-from . import context
-from . import pos
-from . import patient
-from . import location
 from . import activity_extension
 from . import api
-from . import spell
-from . import operations
-from . import adt
-from . import devices
-from . import wizard
 from . import auditing
-from .tests import test_model
+from . import context
+from . import devices
+from . import groups
+from . import location
+from . import operations
+from . import partner
+from . import patient
+from . import pos
+from . import spell
+from . import user
+from . import wizard
+from .models import *
 from .tests import api_demo
+from .tests import test_model
 from .tests.common import nh_clinical_test_utils
