+++ nh_clinical/models/__init__.py
@@ -1,1 +1,2 @@
 from . import user_management
+from .adt import *
